+++ lib/bylines_filters.php
@@ -9,18 +9,17 @@ function get_the_bylines_posts_link() {
     $authors = [];
     $return = '';
     foreach ($bylines as $byline) {
-    //  if ((get_user_by('ID', $byline->ID)) || get_user_by('slug', $byline->slug)) {
-        //
-    //  } else {
-          $authors[] = $byline;
-    //  }
+			$authors[] = $byline;
     }
-    //echo var_dump($authors);
     if((count($authors) > 0)) {
       $return .= '<span style="font-weight: bold;">Authors: </span>';
       $i = 0;
       while ($i < count($authors)) {
-        $return .= '<a href="'.$authors[$i]->link.'">'.$authors[$i]->display_name.'</a>';
+        $return .= '<a href="'
+										. $authors[$i]->link
+										. '">'
+										. $authors[$i]->display_name
+										. '</a>';
         $i++;
         if ($i === (count($authors)-1)) {
           if (count($authors) > 2) {