+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 == Changelog ==
 
+= 1.1.0 =
+* Changed Byline filter to remove errors
+	* probably
+* Fixed links for attachments on single research page
+
 = 1.0.0 =
 * Change to Bylines
 * Add views in plugin